refactored bookmark and results view into new preview view parent component
